refactor(dashboard): Extract channel assignment to shared component
